fix(async): rename module to avoid use of reserved keyword
